move LfContingency from sa to util
